ISBNET/requirements.txt
addict==2.4.0
alembic==1.11.2
amqp==5.2.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
bidict==0.22.1
billiard==4.2.0
blinker==1.6.2
Brotli @ file:///croot/brotli-split_1714483155106/work
build==1.0.3
CacheControl==0.13.1
ccimport==0.3.7
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.3
cumm-cu113==0.2.9
cycler==0.12.1
Cython==3.0.10
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.4.2
dramatiq==1.16.0
dulwich==0.21.7
eventlet==0.33.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-dramatiq==0.6.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.47.2
fsspec==2024.3.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpie==3.2.2
humanfriendly==10.0
humanize==4.9.0
idna==3.4
imagecodecs==2024.6.1
imageio==2.34.1
importlib-metadata==7.0.0
installer==0.7.0
intel-openmp==2021.4.0
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keyring==24.3.0
khandy==0.1.9
kiwisolver==1.4.5
kombu==5.3.5
lark==1.1.9
lazy_loader==0.4
lxml==5.2.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
munch==4.0.0
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.1
open3d==0.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddlepaddle==2.4.2
pandas==2.2.0
parso==0.8.3
pccm==0.3.4
pexpect==4.9.0
pika==1.3.2
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.20.0
plyfile==1.0.3
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pointnet2 @ file:///server/isbnet/ISBNetForPredict/ISBNet/isbnet/pointnet2/dist/pointnet2-0.0.0-cp310-cp310-linux_x86_64.whl#sha256=4e31744b1310e0cafa7699869a5443428e4af3dd41fd76a897147b63b08aed1f
portalocker==2.10.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
PTCAPhase3Pkg-R2024a==24.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pybind11==2.13.1
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyquaternion==0.9.9
pyreadline3==3.4.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2023.3.post1
PyWavelets==1.6.0
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.5.2
rpds-py==0.18.0
scikit-image==0.23.2
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.2
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==0.10.1
six==1.16.0
spconv-cu113==2.1.25
SQLAlchemy==2.0.19
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.5.22
tomli==2.0.1
tomlkit==0.12.3
torch==1.12.1
torch-scatter==2.0.9
torchvision==0.13.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
trove-classifiers==2024.1.31
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2023.4
ultralytics==8.2.50
ultralytics-thop==2.0.0
urllib3==2.0.4
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==2.3.6
widgetsnbextension==4.0.10
wsproto==1.2.0
yapf==0.40.2
zipp==3.17.0
